Remove some low-value, slow tests
The removed tests use real stubs and/or run Python and thus are pretty
slow to run. Most of them either test stubs in straightforward ways
(which is not very useful, since the coverage is bound to be spotty)
or some very basic functionality (which is covered by other tests).

The removals speed up the Python evaluation suite by 10-15%. There are
probably others that could be removed but I tried to be pretty
conservative here.