Always test cache file generation during testcheck
There was already some code for doing this testing during incremental
testing, but it was conditioned on `testcase.output is None`, which
was always false.
It was used as a starting point but when mypy deletes cache files
seems to have changed quite a bit since that code had been written.

This makes running testcheck slower but not vastly so: about 10% in my
ad-hoc testing.

Fixes #4742.